refactor(javm-bench): dedup the two sub-VM recurse benches
sub_vm_recurse.rs and sub_vm_data_recurse.rs were ~95% identical (Built
struct, build_and_publish, invoke, hex_short, criterion driver) and each
re-derived Instance overlays inline. Promote `criterion` to a normal
(target-scoped) dependency — javm-bench is a benchmark crate — so the
whole driver can live in src/lib.rs:

- add `build_sub_vm_top` (using the canonical `Image::data_overlays()`),
  `invoke_sub_vm`, `hex_short`, and the `run_recurse_bench` criterion
  driver to the lib
- collapse each bench file to a blob const + a one-line call (237/209 ->
  ~40 lines), dropping the dead `depth_seed` parameter
- fix two stale "byte-PVM" mentions (it is the PVM2 / RISC-V interpreter)

Both benches pass `cargo bench -- --test`.
